[TASK] Changes to the schematic
Exchanged the output transistor to a MOSFET.

Added some custom adapted symbols and created a gafrc to enable
access to them in gschem.

Added a transistor which shuts down the IR-LED and IR Photodiode
after the circuit got enabled (IR beam interrupted) and self hold
is active. This ensures all devices inside the trap are voltage
free after activation: This could save a mouse some tremor.